move remaining ray stuff
